requirements.txt
streamlit==1.54.0
pyswisseph==2.10.3.2